Optimisation
Set channeling to ai = no because it was already set to 0 for ai_will_do. Spell ai_intervals multiplied by 5 and ai_will_do multiplied by 5 - so we check less frequently but get results on more of those checks. This is only a first pass and we still need a more robust solution.